Move utility methods out of core subpackage
