+++ doc/api/core/misc.rst
@@ -6,5 +6,5 @@ These miscellaneous methods are used to support imaging or data sampling.
 .. autosummary::
    :toctree: _autosummary
 
-   pyvista.core.imaging.sample_function
-   pyvista.core.common_data.perlin_noise
+   pyvista.utilities.common.sample_function
+   pyvista.utilities.common.perlin_noise
