+++ examples/01-filter/image-fft-perlin-noise.py
@@ -9,8 +9,8 @@ This example shows how to apply a Fast Fourier Transform (FFT) to a
 filter.
 
 Here, we demonstrate FFT usage by first generating Perlin noise using
-:func:`pyvista.sample_function() <pyvista.core.imaging.sample_function>` to
-sample :func:`pyvista.perlin_noise <pyvista.core.common_data.perlin_noise>`,
+:func:`pyvista.sample_function() <pyvista.utilities.common.sample_function>` to
+sample :func:`pyvista.perlin_noise <pyvista.utilities.common.perlin_noise>`,
 and then performing FFT of the sampled noise to show the frequency content of
 that noise.
 """
@@ -65,7 +65,7 @@ subset = sampled_fft.extract_subset((0, xdim // 2, 0, ydim // 2, 0, 0))
 # ~~~~~~~~~~~~~~~~~~~~~~~~~
 # Now, plot the noise in the frequency domain. Note how there is more high
 # frequency content in the x direction and this matches the frequencies given
-# to :func:`pyvista.perlin_noise <pyvista.core.common_data.perlin_noise>`.
+# to :func:`pyvista.perlin_noise <pyvista.utilities.common.perlin_noise>`.
 
 # scale to make the plot viewable
 subset['scalars'] = np.abs(subset.active_scalars)
