+++ examples/01-filter/sampling_functions_2d.py
@@ -3,7 +3,7 @@
 
 Sample Function: Perlin Noise in 2D
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-Here we use :func:`pyvista.core.imaging.sample_function` to sample
+Here we use :func:`pyvista.utilities.common.sample_function` to sample
 Perlin noise over a region to generate random terrain.
 
 Perlin noise is atype of gradient noise often used by visual effects
