+++ examples/01-filter/sampling_functions_3d.py
@@ -1,7 +1,7 @@
 """
 Sample Function: Perlin Noise in 3D
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-Here we use :func:`pyvista.core.imaging.sample_function` to sample
+Here we use :func:`pyvista.utilities.common.sample_function` to sample
 Perlin noise over a region to generate random terrain.
 
 Video games like Minecraft use Perlin noise to create terrain.  Here,
