+++ pyvista/core/__init__.py
@@ -21,6 +21,4 @@ from .pointset import (
     ExplicitStructuredGrid,
 )
 from .pyvista_ndarray import pyvista_ndarray
-from .common_data import perlin_noise
-from .imaging import sample_function
 from .cell import CellType
