+++ pyvista/utilities/__init__.py
@@ -1,4 +1,5 @@
 """Utilities routines."""
+from .common import perlin_noise, sample_function
 from .errors import (
     GPUInfo,
     Observer,
